// src/@fake-db/types.ts
export type PDV = {
  id: number
  nom: string
  ville: string
  enseigne: string
  region: string
  localisation: string
  checked: boolean
}